[REF] Core: move params definitions to respective plugins
